requirements.txt
google-generativeai==0.8.3
chromadb==0.5.15
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3
sqlalchemy==2.0.36
pyodbc==5.2.0
python-dotenv==1.0.1
httpx==0.28.1
numpy==1.26.4
sentence-transformers==3.3.1
scikit-learn==1.6.0
python-multipart==0.0.20
jinja2==3.1.4
aiofiles==24.1.0

